Added a KernelElf class that is supposed to encapsulate all the ELF magic only needed for the kernel image (and the kernel modules). Removed the BootstrapInfo class. Unmapped the remaining identity-mapping of the first 2/4MB on x86/64. Removed the MemoryMap class.

+++ src/system/kernel/linker/FileLoader.h
@@ -41,8 +41,8 @@ public:
    */
   virtual const char *lookupSymbol(uintptr_t nAddress, uintptr_t *pSymbolStart) =0;
   
-  virtual uint32_t debugFrameTable() = 0;
-  virtual uint32_t debugFrameTableLength() = 0;
+  virtual uintptr_t debugFrameTable() = 0;
+  virtual uintptr_t debugFrameTableLength() = 0;
 };
 
 #endif
